pyproject.toml
[tool.poetry]
name = "goat_data-preparation"
version = "1.0.0"
description = ""
authors = ["Admin <majk.shkurti@plan4better.de>"]

[tool.poetry.dependencies]
python = "^3.9"
geopandas = "0.9.0"
Cython = "0.29.24"
GeoAlchemy2 = "0.6.3"
geojson = "2.5.0"
geopy = "2.2.0"
psycopg2 = "2.9.2"
pyrobuf = "0.9.3"
pyrosm = "0.6.1"
PyYAML = "5.4.1"
Shapely = "1.8.0"
argparse = "^1.4.0"
sklearn = "^0.0"
joblib = "^1.1.0"
cykhash = "1.0.2"
Fiona = "1.8.20"
h3 = "^3.7.3"
python-decouple = "^3.6"
mkdocs = "^1.2.3"
mkdocs-material = "^8.2.2"
psutil = "^5.9.0"
rich = "^12.4.4"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
debugpy = "^1.4.1"
black = "^21.7b0"
pylint = "^2.12.2"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

